*going insane*

+++ Dashboard.pde
@@ -48,7 +48,7 @@ public class Dashboard
         fill(255);
         textSize(25);
         text("🚩 " + flagsLeft, (width / 4) - 20, height - 32.5);
-        text("⏱ " + minutes + ":" + secondsString, (width * 3 / 4) + 20, height - 32.5);
+        text("⏱ " + (int)minutes + ":" + (int)secondsString, (width * 3 / 4) + 20, height - 32.5);
         textSize(15);
         if (status == 2)
             face.setLabel("😖");
